+++ src/main.cpp
@@ -1096,6 +1096,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE /*hPrevInstance*/, LPSTR lpCmd
         ChiptuneTracker::DrawBrowser(project, uiState, sequencer);
         ChiptuneTracker::DrawPluginsPanel(project, uiState, sequencer);
         ChiptuneTracker::DrawNoteFXPanel(project, uiState, sequencer);
+        ChiptuneTracker::DrawScopes(project, uiState, sequencer);
         ChiptuneTracker::DrawShortcutsPanel(project, uiState, sequencer);
         ChiptuneTracker::DrawCommandPalette(project, uiState, sequencer);